Move common components to core module
