+++ src/core/components/FeatureItem/FeatureItem.js
