+++ src/core/components/FeaturesList/FeaturesList.css
