+++ src/core/components/FeaturesList/FeaturesList.js
