+++ src/core/components/FeaturesList/index.js
