+++ src/core/components/Heading/Heading.js
