+++ src/core/components/Heading/Heading.scss
