+++ src/core/components/Heading/index.js
