+++ src/core/components/Page/index.js
