+++ src/core/components/Text/Text.js
