+++ src/core/components/Text/index.js
