+++ src/dashboard/components/Layout/Layout.js
@@ -3,10 +3,12 @@ import "./Layout.css";
 import React from "react";
 import classNames from "classnames";
 
-import Container from "../Container";
+import Container from "../../../core/components/Container";
 
 export default function Layout({ children, className }) {
   return (
-    <Container className={classNames("Layout", className)}>{children}</Container>
+    <Container className={classNames("Layout", className)}>
+      {children}
+    </Container>
   );
 }
